requirements.txt
#psycopg2>=2.9.0,<3.0.0
pandas>=2.0.0,<3.0.0
gensim>=4.0.0,<5.0.0
numpy>=1.20.0,<2.0.0
psutil>=5.8.0,<6.0.0
ipykernel>=6.0.0,<7.0.0
torch>=2.2.0,<2.7.0
matplotlib>=3.3.0,<4.0.0
tqdm>=4.50.0,<5.0.0
scikit-learn>=0.24.0,<1.4.0
wandb
datasets
transformers
python-dotenv
rank_bm25
nbformat
gensim
requests